Removes XmlSerializer warnings
Seems there is no way to prevent them, and they are there "by design" so
no need to fill up the screen with those warnings when building with
output enabled. There's actually enough of them that it adds about 1s to
the build time. But as we move to some command line driven build tools
this will also help keep the noise down.